refactor: Move shared file content to heleprs
